backend/requirements.txt
Django>=4.2,<5.0
django-ninja>=0.12.0
djangorestframework>=3.12.0
psycopg2-binary>=2.9.0
pydantic>=1.10.0
email-validator>=2.0.0 